% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{evaluaciones}
\alias{evaluaciones}
\title{Modelo de evaluaciones para individuos basado en Val Jonhson}
\usage{
evaluaciones(N, J, E, K, nclases, clases, marcaevaluada, X, y, niter = 15000,
  nburn = 5000, nstep = 100, nadjust = 100, alpha = as.double(0.5),
  lambda = as.double(2), tau0 = as.double(2), sigmaMH = as.double(1))
}
\arguments{
\item{N}{Numero de individuos}

\item{J}{Numero de objetos (marcas, atributos) evaluados}

\item{E}{Valor maximo de la escala de evaluacione 1,..,E}

\item{K}{Cantidad de coeficientes a estimar (nbeta)}

\item{nclases}{Cantidad de clases a estimar}

\item{clases}{Mapa entre las filas de individuos y las clases}

\item{marcaevaluada}{Indicador NxJ de marcas evaluadas por individuo}

\item{X}{Covariables NxJxK}

\item{y}{Evaluaciones realizadas NxJ}

\item{niter}{Numero de iteraciones}

\item{nburn}{Cantidad de iteraciones a quemar del MCMC}

\item{nstep}{Pasos para tomar la simulacion del MCMC}

\item{nadjust}{No recuerdo}

\item{alpha}{Parametros de las previas}

\item{lambda}{Parametros de las previas}

\item{tau0}{Parametros de las previas}

\item{sigmaMH}{Parametros de las previas}
}
\value{
El programa genera archivos de texto con las simulaciones realizadas
}
\description{
Modelo de evaluaciones para individuos basado en Val Jonhson.
Este modelo estima por MCMC los coeficientes de la relación entre las evaluaciones y y las covariables X
El modelo soporta que los individuos no evaluen todos las mismas marcas (por ejemplo cuando no las conocen)
Adicionalmente, el modelo puede estimar los coeficientes en nclases grupos. Si nclases=N el numero de individuos, equivale el modelo jerarquico por individuos,
pero nclases puede ser menor, por ejemplo, coeficientes por estrato.
}
\examples{
## Falta buscar el ejemplo, por ejemplo en los proyectos de modelos de marcas o jerarquizacion de drivers.

}

